Added smoketests for necessary components and moved customConfirm to Utils where more appropriate
