+++ src/components/Confirm/ConfirmDialog.test.js
@@ -1,0 +1,28 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+import renderer from 'react-test-renderer'
+import { MemoryRouter } from 'react-router-dom'
+import ConfirmDialog from './ConfirmDialog'
+
+describe('<ConfirmDialog />', () => {
+  // Smoke test
+  it('renders without crashing', () => {
+    const div = document.createElement('div')
+    ReactDOM.render(
+      <MemoryRouter>
+          <ConfirmDialog />
+      </MemoryRouter>, div)
+    ReactDOM.unmountComponentAtNode(div)
+  })
+  // Snapshot test
+  it('renders the UI as expected', () => {
+    const tree = renderer
+      .create(
+        <MemoryRouter>
+            <ConfirmDialog />
+        </MemoryRouter>
+      )
+      .toJSON()
+    expect(tree).toMatchSnapshot()
+  })
+})
