+++ src/components/GamerTag/GamerTag.test.js
@@ -1,0 +1,31 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+import renderer from 'react-test-renderer'
+import { MemoryRouter } from 'react-router-dom'
+import GamerTag from './GamerTag'
+
+describe('<GamerTag />', () => {
+  const testData = {
+    name: 'TestTag'
+  }
+  // Smoke test
+  it('renders without crashing', () => {
+    const div = document.createElement('div')
+    ReactDOM.render(
+      <MemoryRouter>
+          <GamerTag name={testData.name}/>
+      </MemoryRouter>, div)
+    ReactDOM.unmountComponentAtNode(div)
+  })
+  // Snapshot test
+  it('renders the UI as expected', () => {
+    const tree = renderer
+      .create(
+        <MemoryRouter>
+            <GamerTag name={testData.name}/>
+        </MemoryRouter>
+      )
+      .toJSON()
+    expect(tree).toMatchSnapshot()
+  })
+})
