+++ src/config.js
@@ -1,5 +1,7 @@
 export default {
-  API_ENDPOINT: 'https://project-manhattan.herokuapp.com/api',
+  API_ENDPOINT: 'http://localhost:8000/api',
   TOKEN_KEY: process.env.REACT_APP_TOKEN_KEY,
   SOCKET_CONNECTION: 'https://project-manhattan.herokuapp.com'
 }
+
+// https://project-manhattan.herokuapp.com/api
